refactor: implements interface MenuItem in models header
