+++ src/app/core/models/header/header-account-link.model.ts
@@ -1,8 +1,9 @@
-export class HeaderAccountLink {
+import { MenuItem } from 'primeng/api';
+
+export class HeaderAccountLink implements Partial<MenuItem> {
 	constructor(
 		public label: string,
 		public sublabel: string,
-		public link: string,
 		public icon: string,
 		public command?: () => void,
 	) {}
