+++ src/app/core/models/header/navigation-links.model.ts
@@ -1,7 +1,8 @@
-export class NavigationLink {
+import { MenuItem } from 'primeng/api';
+
+export class NavigationLink implements Partial<MenuItem> {
 	constructor(
 		public label: string,
-		public link: string,
 		public icon: string,
 		public command?: () => void,
 	) {}
