+++ src/app/core/services/header/header-account-link.service.ts
@@ -2,7 +2,7 @@ import { Injectable, WritableSignal, inject } from '@angular/core';
 import { signal } from '@angular/core';
 import { Router } from '@angular/router';
 import { RouteEnum } from '@enum/route.enum';
-import { HeaderAccountLink } from '@models/header-account-link.model';
+import { HeaderAccountLink } from '@models/header/header-account-link.model';
 
 @Injectable({
 	providedIn: 'root',
@@ -14,19 +14,16 @@ export class HeaderAccountLinkService {
 		{
 			label: 'Login',
 			sublabel: 'Account',
-			link: '/auth/login',
 			icon: 'account',
 		},
 		{
 			label: 'Wishlist',
 			sublabel: '3 items',
-			link: '/',
 			icon: 'wishlist',
 		},
 		{
 			label: 'cart',
 			sublabel: '4 items',
-			link: '/',
 			icon: 'cart',
 		},
 	]);
@@ -35,14 +32,12 @@ export class HeaderAccountLinkService {
 		{
 			label: 'Register',
 			sublabel: '',
-			link: '/',
 			icon: '',
 			command: () => this._router.navigateByUrl('/' + RouteEnum.AUTH_REGISTER),
 		},
 		{
 			label: 'Login',
 			sublabel: '',
-			link: '/',
 			icon: '',
 			command: () => this._router.navigateByUrl('/' + RouteEnum.AUTH_LOGIN),
 		},
@@ -52,13 +47,11 @@ export class HeaderAccountLinkService {
 		{
 			label: 'Profile',
 			sublabel: '',
-			link: '/',
 			icon: '',
 		},
 		{
 			label: 'Logout',
 			sublabel: '',
-			link: '/',
 			icon: '',
 		},
 	]);
